crawler/requirements.txt
appdirs==1.4.3
appnope==0.1.0
beautifulsoup4==4.5.3
decorator==4.0.11
html5lib==0.999999999
ipython==5.2.2
ipython-genutils==0.2.0
lxml==3.7.3
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.2.0
PyYAML==3.12
requests==2.14.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
